deleting pfoto not repeats

+++ public/components/Settings/SettingsModel.js
@@ -70,11 +70,14 @@ export class SettingsModel {
                     if (this.settings.user.birthday !== null) {
                         this.settings.user.birthday = this.settings.user.birthday.slice(0, 10);
                     }
-                    Api.getUserPhotoUrl(this.settings.user.id).then(
-                        (image)=>{
-                            this.settings.user.photo = image;
-                        },
-                    );
+                    this.settings.user.photo = '/pics/avatar.png';
+                    if(this.settings.user.image_path){
+                        Api.getUserPhotoUrl(this.settings.user.id).then(
+                            (image)=>{
+                                this.settings.user.photo = image;
+                            },
+                        );
+                    }
                     this.eventBus.emit(SETTINGS_EVENTS.GOT_USER, this.settings);
                 } else {
                     this.eventBus.emit(SETTINGS_EVENTS.UNAUTH);
@@ -88,7 +91,7 @@ export class SettingsModel {
             (response) => {
                 if ( response.status === 200 ) {
                     Api.getUserPhotoUrl(this.settings.user.id).then(
-                        (image) =>{
+                        (image) => {
                             this.settings.user.photo = image;
                             this.eventBus.emit(SETTINGS_EVENTS.PHOTO_UPLOADED, image);
                         },
@@ -99,13 +102,16 @@ export class SettingsModel {
     }
 
     deletePhoto() {
-        Api.deletePhoto().then(
-            (response) => {
-                if ( response.status === 200 ) {
-                    this.eventBus.emit(SETTINGS_EVENTS.PHOTO_UPLOADED, '/pics/avatar.png');
-                }
-            },
-        );
+        if(this.settings.user.photo !== '/pics/avatar.png'){
+            Api.deletePhoto().then(
+                (response) => {
+                    if ( response.status === 200 ) {
+                        this.settings.user.photo = '/pics/avatar.png';
+                        this.eventBus.emit(SETTINGS_EVENTS.PHOTO_UPLOADED, '/pics/avatar.png');
+                    }
+                },
+            );
+        }
     }
 
     logout() {
